backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
scikit-learn
nltk
spacy
pdfminer.six
pandas
python-dotenv
